Move Modals to helper folder
