+++ client/src/App.jsx
@@ -3,7 +3,7 @@ import RelatedItems from './RelatedItems/RelatedItems.jsx';
 import RatingsAndReviews from './RatingsAndReviews/RatingsAndReviews.jsx';
 import ProductOverview from './ProductOverview/ProductOverview.jsx';
 import QuestionsAndAnswers from './QuestionsAndAnswers/QuestionsAndAnswers.jsx';
-import Modal from './QuestionsAndAnswers/AddQuestionModal/Modal.jsx';
+import Modal from './helper/Modals/Modal.jsx';
 
 function App() {
   const [modal, setModal] = useState('');
